Jogadas do computador com níveis de dificuldade

+++ src/main/scala/Settings.scala
@@ -1,1 +1,1 @@
-case class Settings(boardWidth: Int, boardHeight: Int, captureGoal: Int, timeLimitMillis: Long, difficulty: Double)
+case class Settings(boardWidth: Int, boardHeight: Int, captureGoal: Int, timeLimitMillis: Long, difficulty: Int)
